requirements.txt
numpy
PyMuPDF
pillow
matplotlib
docx2pdf; platform_system != "Linux"